A2A_system_low_carbon/requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.36.0
vllm>=0.9.0
accelerate>=0.21.0

# Embeddings and Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
FlagEmbedding>=1.2.10

# Math Verification
math-verify>=0.1.0

# Web Search and Scraping
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
crawl4ai>=0.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.14.0

# Evaluation
evaluate>=0.4.0
scikit-learn>=1.3.0

# Utilities
openai>=1.0.0  # For OpenAI-compatible API
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Jupyter for analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0